Patch from gridcentric to workaround gc-migrate bug.
---------- Forwarded message ----------
From: Adin Scannell <adin@gridcentric.com>
Date: Mon, Aug 5, 2013 at 9:02 AM
Subject: Re: GridCentric BUG
To: Nick Bartos <nick@pistoncloud.com>

Hi Nick,

Wanted to provide an update on the outstanding issues.

We've got a few fixes in progress, but they're not quite ready to go.
I expect that we'll be shipping it to you in the first half of this
week. Apologies for the delay, I understand you'll want to get things
rolled into the image as early as possible for your qa process.

On the migration issue, David has prepared an initial patch for
quantum in order to address the problem (attached). As he notes,
there may exist some corner cases that aren't considered, but we'll be
looking at this when we run our suite against the cluster prior to
shipping.

Let me know if there's anything else that I can do in the meantime.

-Adin

---------- Forwarded message ----------
From: "David Scannell" <dscannell@gridcentric.com>
To: "Adin Scannell" <adin@gridcentric.com>
Cc:

Here is an initial stab at a patch to get migration working with
quantum + multi-host enabled. I have tested it a couple of time
(say ~10) and it works. There might be a chance that it needs some
hardening (catching exceptions, etc.) but on the happy path it
appears to work.

This basically patches the "setup_networks_on_host" function of
the "nova.network.quantumv2.api" to have it update the
"binding:host_id" attribute of the instance's port to the desired
host. This method gets at the beginning of migration to setup the
networks on the destination host. It will also get called in the
event of a roll back to have the network re-setup back on the
original source host.

Thanks,
David.